fix: widen the CUDA stream parameter to intptr_t and forward it for the distance ops (#35)
A `cudaStream_t` is a pointer and therefore 64-bit, but every public
`stream` parameter in this dispatch layer was declared `int`. A real stream
handle passed in from Python is silently truncated to its lower 32 bits and
then reinterpreted as a different, invalid stream. Widen all 128 public
`stream` parameters to `intptr_t`, matching the launcher signatures in
fastfields-cuda-impl.

`dt_euclidean`, `dt_l1` and `dt_mesh` also stopped forwarding the value:
they marked the parameter `/* stream <unused> */` because the underlying
launchers took no stream at all. Now that those launchers accept one, thread
it through the internal `_dt_*` helpers and the dtype dispatch so the caller's
stream reaches the kernel launch.

The three `dt_spline_*` entry points keep `/* stream <unused> */`: the CUDA
spline distance launcher does not exist yet, so there is nothing to forward
to. `posdef`/`resize`/`restrict`/`splinc` likewise still mark it unused here
-- their impl launchers do accept a stream, so wiring this layer up to them
is a real remaining gap, but it is outside the scope of issue #4, which
scopes the forwarding fix to the distance launchers.

Bumps the `impl` submodule pin to cc2155c (the matching cuda-impl change).

Verified by compiling `distance.o` and `posdef.o` with nvcc 12.0 against the
bumped impl. No GPU is available, so this is compile+link verification only.

Refs fastfields/fastfields-lib#4

+++ include/fastfields/api/cuda/reg_field.h
@@ -36,7 +36,7 @@ void field_matvec(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -52,7 +52,7 @@ void field_addmatvec_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -68,7 +68,7 @@ void field_submatvec_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -83,7 +83,7 @@ void field_diag(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -102,7 +102,7 @@ void field_adddiag_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -121,7 +121,7 @@ void field_subdiag_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -138,7 +138,7 @@ void field_kernel(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -157,7 +157,7 @@ void field_addkernel_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -176,7 +176,7 @@ void field_subkernel_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -199,7 +199,7 @@ void field_relax(
           int8_t     bound     = 0,
           int        ndim      = 1,
           int        nb_iter   = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -223,7 +223,7 @@ void field_forward(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -246,7 +246,7 @@ void field_precond(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -265,7 +265,7 @@ void field_precond_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -286,7 +286,7 @@ void field_matvec_rls(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -302,7 +302,7 @@ void field_diag_rls(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -327,7 +327,7 @@ void field_relax_rls(
           int8_t     bound     = 0,
           int        ndim      = 1,
           int        nb_iter   = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 } // namespace cuda
